Trait libra_secure_time::TimeService

pub trait TimeService {
    fn now(&self) -> u64;

    fn sleep(&self, seconds: u64);
}

A generic service for providing time related operations (e.g., returning the current time and sleeping).

Required methods

fn now(&self) -> u64

Returns the current time since the UNIX_EPOCH in seconds as a u64.

fn sleep(&self, seconds: u64)

Sleeps the calling thread for (at least) the specified number of seconds. This call may sleep longer than specified, never less.

Loading content...